requirements.txt
blinker
click
Flask
gunicorn
importlib-metadata
itsdangerous
Jinja2
MarkupSafe
packaging
python-dotenv
Werkzeug
zipp
tensorflow==2.17.0
numpy
tf_keras
pillow
imutils
opencv-python
